requirements.txt
praw
python-dotenv
pandas
nltk
matplotlib

